# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Aries Next API
  description: Aries Next 博客系统 API（由 utoipa 注解自动生成，权威契约即本文件）
  license:
    name: MIT
    identifier: MIT
  version: 0.1.0
servers:
- url: /
paths:
...
  /api/admin/settings/{group}:
    get:
      tags:
      - Admin Settings Groups
      summary: 读取分组设置
      description: 读取分组设置，需要 `settings:manage` 权限。email 组的 `smtp_password`、ai 组的 `api_key` 为 write-only，响应只携带 `*_set` 表示是否已设置。
      operationId: getSettingGroup
      parameters:
      - name: group
        in: path
        description: 分组名：appearance / email / integrations / ai
        required: true
        schema:
          type: string
      responses:
        '200':
          description: 分组设置与乐观锁版本
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingGroupResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无 settings:manage 权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 分组名非法（SETTING_GROUP_NOT_FOUND）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - cookieAuth: []
    put:
      tags:
      - Admin Settings Groups
      summary: 更新分组设置
      description: 更新分组设置，需要 `settings:manage` 权限。乐观锁：`expected_version` 与当前版本不匹配返回 409 `SETTING_VERSION_CONFLICT`，成功后 `version` 递增。appearance / integrations 组为全量覆盖；email / ai 组为部分合并（PATCH 语义）：字段省略或 `null` 保持不变。email 组 `smtp_password`、ai 组 `api_key` 三态：省略或 `null` 保持不变、空字符串清除、非空字符串更新。
      operationId: updateSettingGroup
      parameters:
      - name: group
        in: path
        description: 分组名：appearance / email / integrations / ai
        required: true
        schema:
          type: string
      requestBody:
        description: 分组设置更新体
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingUpdateRequest'
        required: true
      responses:
        '200':
          description: 返回更新后的分组设置
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingGroupResponse'
        '400':
          description: Payload 结构不合法（INVALID_SETTING_PAYLOAD）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无 settings:manage 权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 分组名非法（SETTING_GROUP_NOT_FOUND）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 乐观锁版本冲突（SETTING_VERSION_CONFLICT）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - cookieAuth: []
...
components:
  schemas:
...
    ErrorBody:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
          description: 机器可读错误码，如 `UNAUTHORIZED`、`ARTICLE_NOT_FOUND`。
        details:
          description: 结构化补充信息（如引用计数），大多数错误为空，序列化时省略。
        message:
          type: string
          description: 人类可读错误信息。
    ErrorResponse:
      type: object
      description: |-
        统一错误响应体（与 `http::error::ApiError` 的实际 JSON 形状一致）：
        `{ "error": { "code", "message", "details?" } }`。
        所有端点的 4xx/5xx 响应统一引用本类型。
      required:
      - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
...
    SettingGroupResponse:
      type: object
      title: SettingGroupResponse
      description: 分组设置与乐观锁版本；`settings` 结构按 group 分别为 `AppearanceSettings` / `EmailSettingsView` / `IntegrationSettings` / `AiSettingsView`。
      required:
      - group
      - version
      - updated_at
      - settings
      properties:
        group:
          type: string
          description: 分组名：`appearance` / `email` / `integrations` / `ai`。
        settings:
          description: 分组设置内容，JSON 结构随 group 而不同。
        updated_at:
          type: string
          format: date-time
        version:
          type: integer
          format: int32
          description: 乐观锁版本号，每次更新递增。
    SettingUpdateRequest:
      type: object
      title: UpdateSettingGroupRequest
      description: 分组设置更新体。`settings` 结构按 group 而不同：appearance / integrations 组全量覆盖（`AppearanceSettings` / `IntegrationSettings`）；email / ai 组部分合并（`EmailSettingsPatch` / `AiSettingsPatch`），字段省略或 `null` 保持不变。乐观锁：`expected_version` 与当前版本不匹配返回 409 `SETTING_VERSION_CONFLICT`。
      required:
      - expected_version
      - settings
      properties:
        expected_version:
          type: integer
          format: int32
          description: 乐观锁版本号（先 GET 读取当前 version）。
        settings:
          description: 分组设置内容，JSON 结构随 group 而不同（见 description）。
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: aries_admin_session